projects/stylematch-ai/requirements.txt
requests>=2.32.0
sentence-transformers>=3.0.0
numpy>=1.26.0
pandas>=2.2.0
python-dotenv>=1.0.1
msgpack>=1.0.8
